requirements.txt
httpx==0.22.0
nonebot==1.9.1
nonebot_plugin_apscheduler==0.1.2
nonebot_plugin_chatrecorder==0.1.5
nonebot_plugin_code==0.0.4
nonebot_plugin_fortune==0.4.3
nonebot_plugin_memes==0.2.3.1
nonebot_plugin_petpet==0.2.13
nonebot_plugin_shindan==0.2.2
nonebot_plugin_wordcloud==0.1.0
nonebot_plugin_simplemusic==0.1.1
pydantic==1.9.0
httpx==0.21.3
numpy==1.22.3
opencv_python_headless==4.5.5.64
Pillow==9.1.0
beautifulsoup4==4.10.0
